added Comments and updated Readme.md

+++ budgetapp/forms.py
@@ -2,5 +2,5 @@ from django import forms
 
 class ExpenseForm(forms.Form):
     title = forms.CharField()
-    amount = forms.IntegerField()
+    amount = forms.DecimalField(max_digits=8 , decimal_places=2)
     category = forms.CharField()
